Update EDA-AmeriCorps.rmd
I edited the narrative and reformatted a bit so it looks more like a paper. We may still want to take out the specific questions and write them in as narrative instead. I also added references at the bottom in APA style as required per the instructions. I can't knit it without changing your file path so once you merge and knit a new html, can see if we want to make further changes. I also changed the code folding to hide per her requirements so you should not see the code on the html unless you change the setting the upper right hand corner.